alterando para integer

+++ src/main/java/br/com/giorni/gerenciadororcamento/service/dto/OrcamentoInputDTO.java
@@ -17,7 +17,7 @@ public class OrcamentoInputDTO {
     String observacoes;
     @NonNull
     @JsonProperty("id_servicos")
-    List<Long> idServicos;
+    List<Integer> idServicos;
     @JsonProperty("id_cliente")
     long idCliente;
 }
